pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "toggl.py"
version = "1.0.0a1"

authors = [{name = "Interitio"}]
requires-python = ">= 3.10"

dependencies = [
  "aiohttp",
  "tomllib",
  "attrs",
]

